+++ Sources/KeyboardKit/Views/Callouts/InputCallout.swift
@@ -31,35 +31,37 @@ public struct InputCallout: View {
         self._keyboardContext = ObservedObject(wrappedValue: keyboardContext)
         self.style = style
     }
-    
-    @ObservedObject private var context: InputCalloutContext
-    @ObservedObject private var keyboardContext: KeyboardContext
+
+    @ObservedObject
+    private var context: InputCalloutContext
+
+    @ObservedObject
+    private var keyboardContext: KeyboardContext
 
     private let style: InputCalloutStyle
 
     static let coordinateSpace = InputCalloutContext.coordinateSpace
     
     public var body: some View {
-        VStack(spacing: 0) {
-            callout.offset(y: 1)
-            button
-        }
-        .compositingGroup()
-        .opacity(context.isActive ? 1 : 0)
-        .calloutShadow(style: calloutStyle)
-        .position(x: positionX, y: positionY)
+        callout
+            .transition(.opacity)
+            .opacity(context.isActive ? 1 : 0)
+            .calloutShadow(style: calloutStyle)
+            .position(position)
     }
 }
 
 private extension InputCallout {
-    
-    var button: some View {
-        CalloutButtonArea(
-            frame: buttonFrame,
-            style: calloutStyle)
-    }
-    
+
     var callout: some View {
+        VStack(spacing: 0) {
+            calloutBubble.offset(y: 1)
+            calloutButton
+        }
+        .compositingGroup()
+    }
+
+    var calloutBubble: some View {
         Text(context.input ?? "")
             .font(style.font)
             .frame(minWidth: calloutSize.width, minHeight: calloutSize.height)
@@ -67,7 +69,16 @@ private extension InputCallout {
             .background(calloutStyle.backgroundColor)
             .cornerRadius(cornerRadius)
     }
+    
+    var calloutButton: some View {
+        CalloutButtonArea(
+            frame: buttonFrame,
+            style: calloutStyle)
+    }
 }
+
+
+// MARK: - Private View Properties
 
 private extension InputCallout {
     
@@ -77,9 +88,13 @@ private extension InputCallout {
             dy: buttonInset.height)
     }
     
-    var buttonInset: CGSize { calloutStyle.buttonInset }
+    var buttonInset: CGSize {
+        calloutStyle.buttonInset
+    }
     
-    var buttonSize: CGSize { buttonFrame.size }
+    var buttonSize: CGSize {
+        buttonFrame.size
+    }
     
     var calloutSize: CGSize {
         CGSize(
@@ -102,7 +117,13 @@ private extension InputCallout {
     var cornerRadius: CGFloat {
         shouldEnforceSmallSize ? calloutStyle.buttonCornerRadius : calloutStyle.cornerRadius
     }
-    
+}
+
+
+// MARK: - Private Functionality
+
+private extension InputCallout {
+
     var shouldEnforceSmallSize: Bool {
         #if os(iOS)
         keyboardContext.screenOrientation.isLandscape &&
@@ -113,69 +134,74 @@ private extension InputCallout {
         return false
         #endif
     }
-    
+
+    var position: CGPoint {
+        CGPoint(x: positionX, y: positionY)
+    }
+
     var positionX: CGFloat {
         buttonFrame.origin.x + buttonSize.width/2
     }
-    
+
     var positionY: CGFloat {
         buttonFrame.origin.y + buttonSize.height/2 - calloutSize.height/2
     }
 }
 
+
+// MARK: - Previews
+
 struct InputCallout_Previews: PreviewProvider {
     
-    static let context1 = InputCalloutContext(isEnabled: true)
-    static let context2 = InputCalloutContext(isEnabled: true)
-    
-    static var button: some View {
-        Color.red.frame(width: 40, height: 45)
-    }
-    
-    static var rowItem: some View {
-        Color.yellow.frame(width: 46, height: 51)
-            .overlay(button)
-    }
-    
-    static var rowItemStyle: InputCalloutStyle {
-        var style = InputCalloutStyle.standard
-        style.callout.buttonInset = CGSize(width: 3, height: 3)
-        return style
-    }
-    
-    static var previews: some View {
-        VStack {
-            
-            // Button
-            button.overlay(
-                GeometryReader { geo in
-                    Color.clear.onAppear {
-                        context1.updateInput(for: .character("a"), in: geo)
-                    }
-                }
-            ).inputCallout(
-                context: context1,
-                keyboardContext: .preview,
-                style: .standard)
-            //  style: .preview1)
-            //  style: .preview2)
-            
-            
-            // Row Item
-            
-            rowItem.overlay(
-                GeometryReader { geo in
-                    Color.clear.onAppear {
-                        context2.updateInput(for: .character("a"), in: geo)
-                    }
-                }
-            ).inputCallout(
-                context: context2,
-                keyboardContext: .preview,
-                style: rowItemStyle)
-            //  style: .preview1)
-            //  style: .preview2)
-            
+    struct Preview: View {
+
+        var style: InputCalloutStyle {
+            var style = InputCalloutStyle.standard
+            style.callout.backgroundColor = .blue
+            style.callout.textColor = .white
+            style.callout.buttonInset = CGSize(width: 3, height: 3)
+            return style
         }
+
+        @StateObject
+        var context = InputCalloutContext(isEnabled: true)
+
+        func button(for context: InputCalloutContext) -> some View {
+            GeometryReader { geo in
+                Button(action: { showCallout(for: geo) }) {
+                    Color.red
+                        .cornerRadius(5)
+                }
+            }
+            .frame(width: 40, height: 45)
+            .padding()
+            .background(Color.yellow.cornerRadius(6))
+        }
+
+        func showCallout(for geo: GeometryProxy) {
+            context.updateInput(for: .character("a"), in: geo)
+            context.resetWithDelay()
+        }
+
+        var body: some View {
+            VStack {
+                HStack {
+                    button(for: context)
+                    button(for: context)
+                    button(for: context)
+                }
+                Button("Reset") {
+                    context.reset()
+                }
+            }
+            .inputCallout(
+                context: context,
+                keyboardContext: .preview
+            )
+        }
+    }
+
+    static var previews: some View {
+        Preview()
     }
 }